+++ packages/rest-demo/.gitignore
@@ -1,2 +1,3 @@
 dbs/
-server.js
+dist/
+dist-server/
