+++ packages/rest-demo/package.json
@@ -4,8 +4,8 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "tsc -p ./tsconfig.server.json; node ./server.js",
-    "start": "NODE_ENV=production node ./server.js",
+    "dev": "tsc -p ./tsconfig.server.json; node ./dist-server/server.js",
+    "start": "NODE_ENV=production node ./dist-server/server.js",
     "build": "tsc && vite build",
     "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
   },
