+++ packages/rest-demo/server/DBWrapper.ts
@@ -1,87 +1,10 @@
 import Database from "better-sqlite3";
 import { extensionPath } from "@vlcn.io/crsqlite";
-import express from "express";
-import ViteExpress from "vite-express";
-import * as http from "http";
-import {
-  cryb64,
-  encode,
-  decode,
-  tags,
-  hexToBytes,
-  Change,
-  Changes,
-} from "@vlcn.io/ws-common";
+import { cryb64, Change, Changes } from "@vlcn.io/ws-common";
 import path from "path";
 import fs from "fs";
-
-const PORT = parseInt(process.env.PORT || "8080");
 const DB_FOLDER = "./dbs";
 const SCHEMA_FOLDER = "./src/schemas";
-
-const app = express();
-const server = http.createServer(app);
-
-app.get("/changes/:room", async (req, res) => {
-  const db = await createDb(
-    req.params.room,
-    req.query.schemaName as string,
-    BigInt(req.query.schemaVersion as string)
-  );
-  try {
-    const requestorSiteId = hexToBytes(req.query.requestor as string);
-    const sinceVersion = BigInt(req.query.since as string);
-
-    console.log(
-      `Asked for changes since: ${sinceVersion} requestor: ${requestorSiteId}`
-    );
-
-    const changes = db.getChanges(sinceVersion, requestorSiteId);
-    const encoded = encode({
-      _tag: tags.Changes,
-      changes,
-      sender: db.getId(),
-      since: [sinceVersion, 0],
-    });
-    res.setHeader("Content-Type", "application/octet-stream");
-
-    console.log(`returning ${changes.length} changes`);
-    res.send(encoded);
-  } finally {
-    db.close();
-  }
-});
-
-app.post("/changes/:room", express.raw(), async (req, res) => {
-  const data = new Uint8Array(req.body.buffer);
-
-  const msg = decode(data);
-  if (msg._tag != tags.Changes) {
-    throw new Error(`Expected Changes message but got ${msg._tag}`);
-  }
-  console.log(
-    `Received ${msg.changes.length} changes from ${msg.sender} for ${req.params.room}.
-    Schema ${req.query.schemaName} version ${req.query.schemaVersion}`
-  );
-
-  const db = await createDb(
-    req.params.room,
-    req.query.schemaName as string,
-    BigInt(req.query.schemaVersion as string)
-  );
-  try {
-    db.applyChanges(msg);
-    res.send({ status: "OK" });
-  } finally {
-    db.close();
-  }
-});
-
-server.listen(PORT, () =>
-  console.log("info", `listening on http://localhost:${PORT}!`)
-);
-
-ViteExpress.bind(app, server);
 
 class DBWrapper {
   #db;
@@ -129,7 +52,7 @@ class DBWrapper {
 // the cost of re-constructing it (initializing SQLite, loading the cr-sqlite extension) every request.
 // You should also prepare the statements once and cache them in that world.
 // That is how the websocket and direct-connect servers work.
-async function createDb(
+export async function createDb(
   room: string,
   requestedSchemaName: string,
   requestedVersion: bigint
